chardev: Let chr_sync_read() use unsigned type
Note than tcp_chr_recv() already use size_t.

Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20181012002217.2864-6-philmd@redhat.com>

+++ chardev/char-socket.c
@@ -494,7 +494,7 @@ static gboolean tcp_chr_hup(QIOChannel *channel,
     return G_SOURCE_REMOVE;
 }
 
-static int tcp_chr_sync_read(Chardev *chr, const uint8_t *buf, int len)
+static size_t tcp_chr_sync_read(Chardev *chr, const uint8_t *buf, size_t len)
 {
     SocketChardev *s = SOCKET_CHARDEV(chr);
     int size;
